additional types in NumberPrompt<T>
