+++ libraries/Microsoft.Bot.Builder.Dialogs/Internal/NumberPromptInternal.cs
@@ -68,6 +68,30 @@ namespace Microsoft.Bot.Builder.Dialogs
                         await Validate(context, numberResult);
                     }
                 }
+                else if (typeof(T) == typeof(double))
+                {
+                    if (double.TryParse(result.Resolution["value"].ToString(), out double value))
+                    {
+                        numberResult.Status = PromptStatus.Recognized;
+                        numberResult.Value = (T)(object)value;
+                        numberResult.Text = result.Text;
+                        await Validate(context, numberResult);
+                    }
+                }
+                else if (typeof(T) == typeof(decimal))
+                {
+                    if (decimal.TryParse(result.Resolution["value"].ToString(), out decimal value))
+                    {
+                        numberResult.Status = PromptStatus.Recognized;
+                        numberResult.Value = (T)(object)value;
+                        numberResult.Text = result.Text;
+                        await Validate(context, numberResult);
+                    }
+                }
+                else
+                {
+                    throw new NotSupportedException($"type argument T of type 'typeof(T)' is not supported");
+                }
             }
             return numberResult;
         }
